build: use MODULARIZE option
