+++ CMakeLists.txt
@@ -38,9 +38,9 @@ if (COMPILER_BASENAME STREQUAL emcc)
   add_executable(emscripten-00 src/emscripten-00.cpp)
   target_link_libraries(emscripten-00 PRIVATE ffmpeg json)
   target_compile_options(emscripten-00 PRIVATE "SHELL: -fexceptions")
-  target_link_options(emscripten-00 PRIVATE "SHELL: --bind -s ALLOW_MEMORY_GROWTH=1")
+  target_link_options(emscripten-00 PRIVATE "SHELL: --bind -s ALLOW_MEMORY_GROWTH=1 -s MODULARIZE=1 --minify 0")
 
   add_executable(emscripten-01 src/emscripten-01.cpp)
   target_link_libraries(emscripten-01 PRIVATE ffmpeg json)
-  target_link_options(emscripten-01 PRIVATE "SHELL: --bind -s ALLOW_MEMORY_GROWTH=1")
+  target_link_options(emscripten-01 PRIVATE "SHELL: --bind -s ALLOW_MEMORY_GROWTH=1 -s MODULARIZE=1 --minify 0")
 endif()
